better file interfaces
